requirements.txt
streamlit
pandas
numpy
yfinance
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
Pillow